+++ src/persist-client/src/fetch.rs
@@ -17,6 +17,8 @@ use anyhow::anyhow;
 use differential_dataflow::difference::Semigroup;
 use differential_dataflow::lattice::Lattice;
 use differential_dataflow::trace::Description;
+use prost::Message;
+use serde::{Deserialize, Serialize};
 use timely::progress::{Antichain, Timestamp};
 use timely::PartialOrder;
 use tracing::{debug_span, trace_span, Instrument};
@@ -24,9 +26,10 @@ use tracing::{debug_span, trace_span, Instrument};
 use mz_persist::indexed::encoding::BlobTraceBatchPart;
 use mz_persist::location::{Blob, SeqNo};
 use mz_persist_types::{Codec, Codec64};
+use mz_proto::{ProtoType, RustType};
 
 use crate::error::InvalidUsage;
-pub use crate::internal::encoding::SerdeLeasedBatch;
+use crate::internal::encoding::ProtoLeasedBatch;
 use crate::internal::machine::retry_external;
 use crate::internal::metrics::Metrics;
 use crate::internal::paths::PartialBlobKey;
@@ -490,5 +493,48 @@ where
             lease must have been returned. This batch was not: {:?}",
             self
         );
+    }
+}
+
+/// This represents the serde encoding for [`LeasedBatch`]. We expose the struct
+/// itself (unlike other encodable structs) to attempt to provide stricter drop
+/// semantics on `LeasedBatch`.
+///
+/// For more details see documentation and comments on:
+/// - [`LeasedBatch`]
+/// - From<SerdeLeasedBatch> for LeasedBatch<T>
+/// - From<LeasedBatch<T>> for SerdeLeasedBatch
+#[derive(Debug, Serialize, Deserialize, Clone)]
+pub struct SerdeLeasedBatch(Vec<u8>);
+
+impl<T: Timestamp + Codec64> From<LeasedBatch<T>> for SerdeLeasedBatch {
+    /// Takes a [`LeasedBatch`] into a [`SerdeLeasedBatch`].
+    ///
+    /// Note that this process in non-commutative with `From<SerdeLeasedBatch>
+    /// for LeasedBatch<T>`. The `SerdeLeasedBatch` that this function generates
+    /// inherits the `LeasedBatch`'s' droppability. However, the value generated
+    /// by `From<SerdeLeasedBatch> for LeasedBatch<T>` is never droppable.
+    ///
+    /// For more details, see [`LeasedBatch`]'s documentation.
+    fn from(x: LeasedBatch<T>) -> Self {
+        SerdeLeasedBatch(x.into_proto().encode_to_vec())
+    }
+}
+
+impl<T: Timestamp + Codec64> From<SerdeLeasedBatch> for LeasedBatch<T> {
+    /// Takes a [`SerdeLeasedBatch`] into a [`LeasedBatch`].
+    ///
+    /// Note that this process in non-commutative with `From<LeasedBatch<T>> for
+    /// SerdeLeasedBatch`. The `LeasedBatch` that this function generates is
+    /// never droppable. However, the value generated by `From<LeasedBatch<T>>
+    /// for SerdeLeasedBatch` inherits the `LeasedBatch`'s droppability.
+    ///
+    /// For more details, see [`LeasedBatch`]'s documentation.
+    fn from(x: SerdeLeasedBatch) -> Self {
+        let proto = ProtoLeasedBatch::decode(x.0.as_slice())
+            .expect("internal error: invalid ProtoLeasedBatch");
+        proto
+            .into_rust()
+            .expect("internal error: invalid ProtoLeasedBatch")
     }
 }
