+++ src/persist-client/src/internal/encoding.rs
@@ -12,24 +12,28 @@ use std::marker::PhantomData;
 
 use differential_dataflow::lattice::Lattice;
 use differential_dataflow::trace::Description;
-use mz_persist_types::{Codec, Codec64};
-use mz_proto::{IntoRustIfSome, ProtoType, RustType, TryFromProtoError};
 use prost::Message;
 use semver::Version;
-use serde::{Deserialize, Serialize};
 use timely::progress::{Antichain, Timestamp};
 use timely::PartialOrder;
 use uuid::Uuid;
+
+use mz_persist_types::{Codec, Codec64};
+use mz_proto::{IntoRustIfSome, ProtoType, RustType, TryFromProtoError};
 
 use crate::error::CodecMismatch;
 use crate::fetch::{LeaseLifeCycle, LeasedBatch, LeasedBatchMetadata};
 use crate::internal::paths::PartialBlobKey;
 use crate::internal::state::{proto_lease_life_cycle, proto_leased_batch_metadata};
 use crate::internal::state::{
-    HollowBatch, ProtoHollowBatch, ProtoLeaseLifeCycle, ProtoLeasedBatch, ProtoLeasedBatchMetadata,
-    ProtoReaderState, ProtoStateRollup, ProtoTrace, ProtoU64Antichain, ProtoU64Description,
-    ProtoWriterState, ReaderState, State, StateCollections, WriterState,
+    HollowBatch, ProtoHollowBatch, ProtoLeaseLifeCycle, ProtoLeasedBatchMetadata, ProtoReaderState,
+    ProtoStateRollup, ProtoTrace, ProtoU64Antichain, ProtoU64Description, ProtoWriterState,
+    ReaderState, State, StateCollections, WriterState,
 };
+
+// This has to be exposed to [`crate::fetch::SerdeLeasedBatch`], which we in
+// turn want to make publicly accessible as part of the persist API.
+pub(crate) use crate::internal::state::ProtoLeasedBatch;
 
 use crate::internal::trace::Trace;
 use crate::read::ReaderId;
@@ -485,50 +489,8 @@ impl RustType<ProtoLeaseLifeCycle> for LeaseLifeCycle {
     }
 }
 
-/// This represents the serde encoding for [`LeasedBatch`]. We expose the struct
-/// itself (unlike other encodable structs) to attempt to provide stricter drop
-/// semantics on `LeasedBatch`.
-///
-/// For more details see documentation and comments on:
-/// - [`LeasedBatch`]
-/// - From<SerdeLeasedBatch> for LeasedBatch<T>
-/// - From<LeasedBatch<T>> for SerdeLeasedBatch
-#[derive(Debug, Serialize, Deserialize, Clone)]
-pub struct SerdeLeasedBatch(Vec<u8>);
-
-impl<T: Timestamp + Codec64> From<LeasedBatch<T>> for SerdeLeasedBatch {
-    /// Takes a [`LeasedBatch`] into a [`SerdeLeasedBatch`].
-    ///
-    /// Note that this process in non-commutative with `From<SerdeLeasedBatch>
-    /// for LeasedBatch<T>`. The `SerdeLeasedBatch` that this function generates
-    /// inherits the `LeasedBatch`'s' droppability. However, the value generated
-    /// by `From<SerdeLeasedBatch> for LeasedBatch<T>` is never droppable.
-    ///
-    /// For more details, see [`LeasedBatch`]'s documentation.
-    fn from(x: LeasedBatch<T>) -> Self {
-        SerdeLeasedBatch(x.into_proto().encode_to_vec())
-    }
-}
-
-impl<T: Timestamp + Codec64> From<SerdeLeasedBatch> for LeasedBatch<T> {
-    /// Takes a [`SerdeLeasedBatch`] into a [`LeasedBatch`].
-    ///
-    /// Note that this process in non-commutative with `From<LeasedBatch<T>> for
-    /// SerdeLeasedBatch`. The `LeasedBatch` that this function generates is
-    /// never droppable. However, the value generated by `From<LeasedBatch<T>>
-    /// for SerdeLeasedBatch` inherits the `LeasedBatch`'s droppability.
-    ///
-    /// For more details, see [`LeasedBatch`]'s documentation.
-    fn from(x: SerdeLeasedBatch) -> Self {
-        let proto = ProtoLeasedBatch::decode(x.0.as_slice())
-            .expect("internal error: invalid ProtoLeasedBatch");
-        proto
-            .into_rust()
-            .expect("internal error: invalid ProtoLeasedBatch")
-    }
-}
-
 impl<T: Timestamp + Codec64> RustType<ProtoLeasedBatch> for LeasedBatch<T> {
+    /// n.b. this is used with [`crate::fetch::SerdeLeasedBatch`].
     fn into_proto(&self) -> ProtoLeasedBatch {
         ProtoLeasedBatch {
             shard_id: self.shard_id.into_proto(),
@@ -539,6 +501,7 @@ impl<T: Timestamp + Codec64> RustType<ProtoLeasedBatch> for LeasedBatch<T> {
         }
     }
 
+    /// n.b. this is used with [`crate::fetch::SerdeLeasedBatch`].
     fn from_proto(proto: ProtoLeasedBatch) -> Result<Self, TryFromProtoError> {
         Ok(LeasedBatch::new(
             proto.shard_id.into_rust()?,
